OMAP4: DSS: Clear BURSTTYPE bit when TILER is not used
A kernel L3 error was found in a case where a hardware overlay
was used with both a Tiler buffer and a linear buffer. The error
occurs when a linear buffer is attached to a hardware overlay
after a Tiler buffer had been attached.
BURSTTYPE bit was not being reset back to 0 for subsequent
non-Tiler cases.

Change-Id: Iaa7ab0de8f97356af859307155a191f38eaf5709
Signed-off-by: Steven Fischer <steven.fischer@motorola.com>
Signed-off-by: Roberto Granados Dorado <x0095451@ti.com>