OMAP4: HSI: Fixed spurious RX interrupt in DMA mode
Due to the polling mechanism, the HSR_DATAAVAILABLE interrupt was re-enabled after DMA transfer
was complete, even if no polling was ongoing. This caused the RX interrupt processing function to
be called, and triggered some spurious FIFO not empty warnings.

Also added an API (hsi_unpoll) to cancel a pending polling request.

Change-Id: Ic98da3cbe72dd073084f997a598ec6f79805b321
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>