syslink: ipu_pm: restore IPU if a message is pending
Check the mailbox and call restore after saving the
IPU context to allow IPU to attend any message that
arrived while it was saving the context. The saving
is needed since IPU already disabled the interruption
in it's side and a reset is needed.

Also increment the waiting time for IPU to finish
the saving context and be ready to be hibernated since
ipu_pm was not giving enough time to finish that task
and was aborting hibernation.

IPU needs to use two flags in share memory to check
if a message arrived in the middle of its own save
context these flags needs to be reflected also in the
ipu_pm side structs.

Change-Id: Ie320a45a6ba5f7954c082e939879fc902a50f050
Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>