OMAP2+: hwmod: Add smart-standby wakeup support
The commit c138b539f02170a6cb399867d930a1568c365f91 was adding
the wakeup support for new OMAP4 IPs. This support is incomplete for
busmaster IPs that need as well to use smart-standby with wakeup.

This new standbymode is suported on HSI and USB_HOST_FS for the moment.

Add the new MSTANDBY_SMART_WKUP flag to mark the IPs that support this
capability.

Enable this new mode when applicable in _enable_wakeup, _disable_wakeup,
_sysc_enable and _sysc_idle.

The omap_hwmod_44xx_data.c will have to be updated to add this new flag.

Change-Id: I14935dcd0544a54f7684120932aa6e8b462d97d3
Signed-off-by: Benoit Cousson <b-cousson@ti.com>
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>
Reported-by: Djamil Elaidi <d-elaidi@ti.com>
Acked-by: Benoit Cousson <b-cousson@ti.com>
Cc: Paul Walmsley <paul@pwsan.com>
Cc: Kevin Hilman <khilman@deeprootsystems.com>