OMAP4: Adding SMP bit handling in the offline/resume paths
In this patch the handling of the SMP bit in Public is added
in the offline/resume paths for all CPUs.
This functionality is available in HS devices only, and
requires PPA 1.0 or greater.

Also local macros are introduced to facilitate the calling of
repetitive portions of code and to improve readability.

Change-Id: I5a329c121ee913bff9d6938111eee2b1edc7ef10
Signed-off-by: Carlos Leija <cileija@ti.com>
Signed-off-by: Jean-Luc Rigaud <jl-rigaud@ti.com>
Signed-off-by: Dan Murphy <dmurphy@ti.com>