OMAP3+: xSi: Removed RX buffer cleaning on read cancel
When a read cancel is made, it should affect only the programmed DMA, not the data
which is already in the RX buffer (FIFO head), as this migth be useful data for the next bunch of frames.
If there is a need to flush the RX FIFO, then usage of the ioctl HSI_IOCTL_FLUSH_RX is
recommended.

Additionally, this attempt to flush the RX buffer was effective only for SSI, for which the register is RW.
For HSI this register is RO, so this code has no effect.

Change-Id: I0dbb7e9df249f6ae683c6672b953439863590f55
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>