OMAP4: HDMI: Rework ACR workaround for configurable CTS interval
Notify event to start/stop the ACR workaround needs to be sent
once the sample HDMI DAI hw_params are known. This is needed as
the CTS interval depends on both the pixel clock and the audio
sample rate. Functionality for dynamic configuration of the audio
sample rate will be added later.

Also, relinquish CPU after sending the the notify event. The task
will wake up when the event ACK arrives. A timeout is used in
case SYSM3 does not ACK the notify event.

Change-Id: Idcab8c4814a121a56e05fc2a41b870b77765c521
Signed-off-by: Ricardo Neri <ricardo.neri@ti.com>