OMAP3: V4L2: Fixed to overlay's start position not to be modified by rotation degrees.
Start position of overlay shouldn't be modified by rotation degrees.

In OMAP3, it did swap start position (info->pos_x and info->pox_y) in case of
rotation is 90 or 270. In Android, this causes mess playback screen when the
video clip has rotation info, 90 or 270.

In OMAP4, it did not swap the start position. I think they already found
above bug in their arch and fixed only for OMAP4 because of not sure how
the overlay works in other arch.

Change-Id: I879c7653bcbd5d8f24488154ddae395b527a62ae
Signed-off-by: Homin Lee <homin.lee@ti.com>