OMAP4: HSI: Fix for undetected CAWAKE transitions in OFF mode
Description of the issue:
When modem toggles the CAWAKE line while OMAP is in OFF mode, the HSI tasklet
responsible to handle the CAWAKE line was sometime not called. This led to
not up-to-date cawake mirror variable, and thus a disconnect between real ACWAKE
(physical) line, and internal acwake variable. This disconnect produces some
missing CAWAKE events like :
"CAWAKE race is detected: HI -> LOW -> HI" or
"CAWAKE race is detected: LOW -> HI -> LOW".

Cause of the issue:
when a CAWAKE event is detected from SCM padconf, we schedule the HSI tasklet
only if HSI clocks are disabled, which is not good because :
* at this point of time, HSI clocks are necessarily OFF
* we __need__ to schedule the tasklet to update acwake mirror variable and calls
  callbacks from upper layers

The root cause of the issue is the unpredictable state of HSI clocks when waking up.
This will be fixed in another commit, as it has an impact on functional clocks and is
not only related to OFF mode.

Change-Id: I6227876216369b4bbbfb12fc3d86e3b53ce1416a
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>