OMAP3+: xSI: Fixed condition to disable HSI Fclock
Description of the issue :
when HSI driver doesn't need the HSI Fclock anymore and attempts to disable them,
it succeeds only if HSI HW is not in use (no write pending) and if CAWAKE and
ACWAKE are low.
But the fact that ACWAKE is high shall not prevent to cut the HSI
Fclocks, espcecially since HSI Fclocks are enabled/disabled dynamically on an a
need basis.

Change-Id: I501d3ae945b201a124d234cde34a0e3c7516dfa2
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>