ASoC: HDMI: Shut off audio only when HDMI is active
In order to avoid using HDMI-DSS framework when
HDMI is OFF. HDMI-Audio power is shut off when
HDMI_EVENT_POWERON event is triggered; however, if system
goes to suspend, DAI shutdown was trying to disable
HDMI-Audio resources after DSS clocks were cutoff.
In order to avoid this scenario, HDMI-Audio power
is forced to be released only when HDMI is still active.

This schema is temporal, once HPD events are reliable,
HDMI-Audio resources will be released base on these
notifications.

Change-Id: I73c53259779e283773123514da2ad8d17ba03675
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Axel Castaneda Gonzalez <x0055901@ti.com>