usb: usb-ehci: reset PHY only if port 0 is in PHY mode
Description# Due to a hardware issue when port 0 is in PHY mode
PHY needs to be reseted. If usbhs_phy_save flag is set at suspend
time PHY will be reset (usbhs_phy_safe) at resume time. This patch
sets the usbhs_phy_save flag only if port 0 is in PHY mode, in all
other cases the reset is not performed.

Change-Id: Iad8defffeac162a211f594773ee64d40b2a7b612
Signed-off-by: Catalin Enache <catalin.enache@ti.com>