OMAP4: DSS2: HDMI: Fix for HDMI notification to observer during HDMI_FIRST_HPD
if (hdmi_connected = false || hdmi_power != HDMI_POWER_FULL)
HDMI_FIRST_HPD should be ignored and no connect notification should be sent
to hdmi_obvserver. HDMI EDID read fails here and sending a notification
to hdmi_observer results in overlay interprepting the default VESA:VGA
timings as new panel timings and tries to configure the overlay accordingly.
This causes the overlay configuration to fail for TV manager if the
panel timings are smaller than the overlay then setmanager fails for TV
defaulting to LCD manager. This will cause both the overlays to be sent
to the LCD resulting in garbled display.

This happens only when trying to connect / disconnect the cable and
the connection is loose causing multiple connects / disconnects / first hpd
interrupts to be sent.

Change-Id: I2f9b876f697feaa730de0400461c6071f05f7e29
Signed-off-by: srinivas pulukuru <srinivas.pulukuru@ti.com>