syslink: ipu_pm: change target power state for IVAHD
Change the targeted next power state for IVAHD to Inactive
per the design. When the IVAHD state transition control mode
is set to HW_AUTO when the codec starts up, this must be
configured to On or Inactive or it will crash when the end
of the frame processing is done.

Change-Id: Ib61649f1e641142b786951bd8f80d5e0b162b30f
Signed-off-by: Paul Hunt <hunt@ti.com>