OMAP: USB Host: Manage HCD_HW_ACCESSIBLE flag during suspend/resume
The HCD_HW_ACCESSIBLE flag is used to indicate whether the USB host
controller is accessible and prevent the USB host driver from
attempting before the USB controller is ready. This flag is set
during the USB bus resume and cleared during the USB bus suspend.

When clearing this flag it is also necessary to disable the USB
host controller interrupt and re-enable the interrupt after the
flag is set. If the interrupt is not disabled, then occassionally
we see a race condition where the USB host interrupt occurs before
we have set the HCD_HW_ACCESSIBLE flag and this prevents the
interrupt from being serviced correctly and the kernel disables the
interrupt. So we need to ensure that the USB host controller
interrupt is not enabled until this flag has been set.

Change-Id: Ibd19f0564b41f1661523411b859757186b70a5ba
Signed-off-by: Jon Hunter <jon-hunter@ti.com>