OMAP3+: HSI/SSI: adding return status in R/W cancel function
When calling hsi_read_cancel() or hsi_write_cancel() function to cancel a Read or Write, the
caller doesn't know if operation was already finished or not.
But in some cases, caller may need to know if it was okay to cancel the
operation, or if due to a race condition, the operation has complete at the
same time as the cancel operation has been decided.

R/W cancel now return 0 if nothing was cancelled, -ECANCELED if cancel succeeded.

Change-Id: I013226a913524b738e4f249828a1b81ae1e43cb3
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>