CMA3000: Program mode correctly.
The mode register on the CMA3000 is on bits
3:1, Hence writing directly the mode without
taking into account bit 0 is wrong.

The set_mode API, takes care of shifting the
bits, and setting the mode correctly.

Also, If no IRQ is available from the data structure,
we disable the IRQ on the CMA, using bit 0 of
the CTRL register.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Change-Id: Iddb93544b14cbfd933cbf3b2f5999254811195fa