syslink: ipu_pm: enable/disable wd in runtime
Add the possibility to enable and disable in
runtime the wd feature.

In order to disable watchdog timer the flag wdt_allowed
should be set to 0 and wd will be disabled.

Set the flag back to 1 to re-enable the wd feature.

This flag is located in the member pm_flags of the sms struct
and it is the 7 bit X0000000b.

Also changing the name of the pm_flags members to match
the kernel code convention.

Change-Id: I6dde896886784a8858dc42a5ceb480f43e3f12b9
Signed-off-by: Miguel Vadillo <vadillo@ti.com>