OMAP4: clock: manage LP mode through work queue
Defers entry into DPLL cascading mode via a work queue.  Default delay is 1
second.  This is needed because it is invalid to enter into DPLL cascading
with both CPU cores online.

The work-around for this is to kick off delayed work one second after the
ASoC layer requests to enter into this mode.  Most of the time this is
enough to allow CPU1 to go offline due to inactivity.  If CPU1 remains
online due to higher-than-normal system activity then the 'check' function
will schedule itself to be re-run in one second later to check again if CPU1
is offline or not.  This will continue until either CPU1 goes offline or
ASoC requests for DPLL cascade to be exited.

Change-Id: I9c5ae6c7d24d2000307c81d46f6a1d94889a3f1c
Signed-off-by: Mike Turquette <mturquette@ti.com>