OMAP4: HSI: Fix for HSI_IOCTL_GET_CAWAKE not working in HSI
The ioctl to get the CAWAKE line status was not working in HSI mode, since there was a test on the
GPIO used for this line. But it is only SSI which uses a GPIO to get the CAWAKE status, not HSI.
HSI IP embeds a CAWAKE status bit and interrupt.
As a consequence, the cawake_gpio value was -1 for HSI, and prevented the use of the ioctl.

Change-Id: I1f58f9d1d422e73a73bef41e44153654779afff9
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>