omap4: remove NEON support from es2.2 and below
There is a silicon bug in Neon on omap4 ES2.2 and below
versions. So make sure NEON is not reported as a feature.

The bug is linked to conditional Store and Load multiple in Neon
that can lead to a processor deadlock
(ARM Errata [754319] and [754320])

%cat /proc/cpuinfo
will not report "neon" for es2.2 and below silicons of omap4
will continue to report "neon" on es2.3 and above omap4 revs

Change-Id: Ib14e061cea649ac9c96e04a2d23f53dab87a16e3
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>