OMAP4: clock: Support divider selection for auxclks
Adaptation for k35 based on:

On OMAP4 the auxclk nodes (part of SCRM) support both
divider as well as parent selection.
Supporting this requires splitting the existing nodes
(which support only parent selection) into
two nodes, one for parent and another for
divider selection.
The nodes for parent selection are named
auxclk*_src_ck and the ones for divider
selection as auxclk*_ck

Change-Id: I44e1c5b15a7842bac475ccdc5897de4d1d949561
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: Miguel Vadillo <vadillo@ti.com>