syslink: ipu_pm: use new api to set auxclks rate
Use clk_set_rate api to set the rate of an auxclk
instead of hard coding the value of the divisor
directly in the register.

Also change the M3 rate to 192 Mhz to be able to set
the aux_clk to 24 Mhz since the M3 is set as the parent
of all aux_clks.

Removing a Hack related with auxclks. Not needed
anymore since MM is now properly requesting the
auxiliary clocks via our APIs when they are needed.

Change-Id: I566a4b8fec58894a272dbd226d3a756f92c006e3
Signed-off-by: Miguel Vadillo <vadillo@ti.com>