OMAP4: DSS: HDMI: do not zero timings
Description# Do not zero panel timings, as they are used by omap_vout
V4L2 driver.
Previously HDMI driver was resetting the panel timings to zero incase of
disconnect and powering down PHY scenarios. But because of multiple HPD events
are possible in the system, there were chances wherein user received HDMI
attach event but the timings were already set to zero by a follow-on detach
event.
So now we keep the timings intact on detach.

Change-Id: Ic6f6ef806260865fa271ed7cc021992b0bf72dd7
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>