SYNTM12XX Touch screen: Add controller number to the board data
This commit will add a controller number to the board data.

This is added so that the controllers can register and report
data as independent controllers.  This will remove the need to
loop through all the controllers to register and report the data.

In addition changed the input dev name to a more Linux friendly
name.

Change-Id: I8ed61488d8b2a1fb381ad8fa1ec57cf0bb35c589
Signed-off-by: Dan Murphy <dmurphy@ti.com>