USB: Reinitlialize LDO on irq.
Beacause LDO setting are lost on suspend,
we need to re-initialize them on a irq event.
if we are shuting down, on a cable detach,
we can set LDO settings for low power.

This patch also adds a 1 second wake lock
on a cable detach, to prevent inmediate suspend
uppon a cable detach.

Change-Id: I2199b8efdd755b4f2a79895eee3c614356436b54
Signed-off-by: Axel Haslam <axelhaslam@ti.com>