OMAP4: HSI: Fix for early DMA interrupt without data transfer
When a DMA R/W cancel is made (hsi_read_cancel() or hsi_write_cancel()), we noticed some
data corruption the next time the cancelled DMA channel is re-used again.

This is occuring because when the DMA read or write is cancelled, HSI driver does not read
DMA_CSR register (clear on read register). Thus the events inside this register could trigger
a new interrupt the next time it is enabled.
So when the DMA channel is re-used again (possibly much later), the DMA interrupt is triggered
as soon as we enable the DMA channel, without any data being transfered.
This gives the impression of a data corruption.

Solution is to add a reading of this register after the channel is disabled (CCR_ENABLE = 0) and before the interrupts status is cleared.

Change-Id: Icb560c63cf99308baf05865636ce6144a2874965
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>