OMAP4 DSS HDMI: Fix EDID read in infinite loop in case of EDID read failure
In case of inifinite looping of EDID read when EDID read fails or sees
VESA:VGA as the timing on the first read attempt the following changes
have been made.
1. The delayed_work_queue causing the re-read of EDID has been removed
2. HDMI mode (CEA/VESA) selection has been changed to
OMAP_HDMI_TIMINGS_VESA_START
3. When FIRST HPD is recieved before Physical connect in case of loose
connection of the cable we have seen edid read failure and infinite
looping due to interrupts not being cleared. This patch also fixes this issue.

Change-Id: I08b0bf746f416ebec71442efe1f26423aa389294
Signed-off-by: Fabrice Olivero <f-olivero@ti.com>
Signed-off-by: Srinivas Pulukuru <srinivas.pulukuru@ti.com>