OMAP3: OPP structure and function changes
Need to put struct omap_opp into a header file as it is also used by the
dspbridge.

When populating an opp strucuture, we also need to record the dev_opp (was
already part of the omap_opp structure, it just wasn't being filled in).
If not done, later calls to opp_enable() will find the opp, but will not
increase the enabled count:
	if (!opp->enabled && opp->dev_opp)
                opp->dev_opp->enabled_opp_count++;
And calls to opp_init_cpufreq_table() will not obtain all enabled entries.

Change-Id: Ic14ff26bec176611efc08742bef6be21de3e52ff
Signed-off-by: Steve Schefter <sschefter@ti.com>