OMAP3+: voltage: remove hardcoded initial voltage
Blindly setting a 1.2V setting in the initial structure may not even
match the default voltages stored in the voltage table which are
supported for the domain. For example, OMAP3430 core domain does not
use 1.2V and ends up generating a warning on the first transition.

Grab the nominal voltage and use it instead

Change-Id: Ie98d7ef2304a63f9986bcb11aa8eee101fe70724
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: jack.miller <jack.miller@ti.com>