OMAP3+: voltage: add transdone APIs
Transdone event in Voltage processor gives us fine grained status on
the current status of the voltage communication with the PMIC.
Unfortunately, IRQ generation by VP is based on the start of the
transmission to VC from VP, not at the end (or the completion of
the voltage setting). Hence any users of voltage layer who need
to know fine grained information such as confirmation if the voltage
is actually send to PMIC, needs to depend on this status.

Change-Id: I6369eace4aaed22607115d874e0b19174ed2ef25
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: jack.miller <jack.miller@ti.com>