OMAP3+: SR: make notify independent of class
Interrupt notification mechanism of SmartReflex can be used by the
choice of implementation of the class driver. For example, Class 2 and
Class 1.5 of SmartReflex can both use the interrupt notification to
identify the transition of voltage or other events.

Hence, the actual class does not matter for notifier. Let the class
driver's handling decide how it should be used. SmartReflex driver
should provide just the primitives.

Change-Id: Ic7d5af38395a92c73480071b52ba072948745813
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: jack.miller <jack.miller@ti.com>