OMAP3+: SR: disable interrupt by default
We will enable and disable interrupt on a need basis in the class
driver. We need to keep the IRQ disabled by default else the
forceupdate or vcbypass events could trigger events that we don't
need/expect to handle.

This is a preparation for SmartReflex AVS class drivers such as
class 2 and class 1.5 which would need to use interrupts. Existing
SmartReflex AVS class 3 driver does not require to use interrupts
and is not impacted by this change.

Change-Id: I01f46695c7d1576fe27e3a6aed87e919270c9419
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: jack.miller <jack.miller@ti.com>