OMAP3+: SR: introduce notifier_control
We need some mechanism from class drivers to control when notifiers
should be triggered and when not, currently we have none, which makes
Class driver usage of the interrupt events almost impossible.

Introduce an SmartReflex driver API for doing the same. This is useful
for SmartReflex AVS class 1.5 or 2 drivers.

Change-Id: I48c9f14ded848b9a08736596ee9f9d7cd3c0ae27
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: jack.miller <jack.miller@ti.com>