OMAP4: HSI: Remove HSI wakeup event when HSI not muxed
When using a platform with no modem, the bootarg parameter "board-4430sdp.modem_ipc=hsi" shall not
be set. But as the omap_hsi driver is built statically (default config is with modem), it is always
reacting to the external inputs, especially to CAWAKE line which can trigger a wakeup from OFF mode.
But in non modem boards, this CAWAKE line is in fact muxed to usbb1_ulpitll_clk, so a transition on this
line can trigger a false wakeup.

So I added a security to check if this PADCONF (offset 0x0C0) is muxed to CAWAKE before :
- enabling the PADCONF wakeup event
- reacting to a PADCONF wakeup event and kick the HSI tasklet

Change-Id: I674d75d1a41704fb3a6eadb0965779f417fa91bd
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>