ASoC: sdp4430: Add vaux2 regulator for AV Switch
An AV switch is used in Blaze/SDP4430 to switch headset jack pin 3
between headset microphone and video output. AV switch is then a
consumer of VAUX2 supply. The corresponding consumer-supply is added
to have VAUX2 supply when headset microphone is in use.

This is considered non-essential, so if the driver is unable to
acquire the VAUX2 supply, it will gracefully continue.

On the 2.6.35 branch, this is the similar to:
    commit: 2b3c5945463fe470b240eee8643cad4db48dae66
    Change-Id: Ibb0264926ca599e8eae99ef0b747e8e08a4e316b

Change-Id: I0cc45977b2e771544b77b13eeaa51ba7987bb382
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>