ASoC: ABE HAL: ABE HAL 9.30 voice recording during BT call
If you are in Bluetooth voice call the ABE operates at OPP50.
If you start in the same time the voice recoding, the ABE
needs to run at OPP100.
In this case the BT ASRC is enable and generate some noise.
This patch fixes this problem in the ABE firmware.

Change-Id: I037a7fe6a6a0b490484544e1535a72a2e098a3fe
Signed-off-by: Francois Mazard <f-mazard@ti.com>