Proximity: This is a work around for the proximity sensor
This is a work around for the proximity sensor since the
sensor was placed in the wrong GPIO bank.  The proximity needs
to be able to wake the system and the GPIO should have been placed
in the wake up domain.  If the prox GPIO is not placed in the
wake up domain the GPIO module will miss the proximity edge and thereby
not calling the interrupt handler.

This patch will introduce a polling mechanism since level interrupts will
not work very well with this type of binary proximity sensor.

When the proximity sensor is turned on the prox GPIO level will be read and reported
When the proximity sensor is off then no polling will take place.

Maintained the ability to use IRQs if the sensor GPIO is in the wakeup domain

Change-Id: I64b1aba52a0be346fedf947d91f26c738fd918d4
Signed-off-by: Dan Murphy <dmurphy@ti.com>